Add professional branding and fix all compiler warnings
Branding improvements:
- Add copyright headers to all source files (Eduardo Stern)
- Update LICENSE files with full name attribution
- Add "By Eduardo Stern" subtitle to README and website
- Update website footers with author credit
- Maintain clean "FluxParser" main brand (no possessive)

Code quality fixes:
- Fix all float/double type mismatches in test files
- Replace fabsf() with fabs() for double precision
- Remove unused variables (interactive, random_seeded_local)
- Add (void) suppressions for unused placeholder parameters
- Clean build: zero warnings with -Wall -Wextra

Website improvements:
- Fix incomplete code example in hero section
- Add complete, runnable examples with expected outputs
- Improve code snippet clarity and correctness

Result: Production-ready codebase with professional attribution
and clean compilation (zero warnings).

+++ example_usage.c
@@ -11,7 +11,7 @@ int main() {
     /* Example 1: Using default single-letter variables (a-z) */
     printf("Example 1: Default single-letter variables\n");
     {
-        float values[] = {10.0, 20.0, 5.0};  /* a=10, b=20, c=5 */
+        double values[] = {10.0, 20.0, 5.0};  /* a=10, b=20, c=5 */
 
         VarContext ctx = {
             .values = values,
@@ -20,14 +20,14 @@ int main() {
             .mapping_count = 0
         };
 
-        float result = parse_expression_with_vars("(a + b) * c", &ctx);
+        double result = parse_expression_with_vars("(a + b) * c", &ctx);
         printf("  (a + b) * c with a=10, b=20, c=5 => %.2f\n", result);
     }
 
     /* Example 2: Custom variable names */
     printf("\nExample 2: Custom variable names\n");
     {
-        float values[] = {100.0, 75.0, 0.15};
+        double values[] = {100.0, 75.0, 0.15};
 
         VarMapping mappings[] = {
             {"PRICE", 0},
@@ -42,10 +42,10 @@ int main() {
             .mapping_count = 3
         };
 
-        float subtotal = parse_expression_with_vars("PRICE * QUANTITY", &ctx);
+        double subtotal = parse_expression_with_vars("PRICE * QUANTITY", &ctx);
         printf("  Subtotal: %.2f\n", subtotal);
 
-        float total = parse_expression_with_vars("PRICE * QUANTITY * (1 + TAXRATE)", &ctx);
+        double total = parse_expression_with_vars("PRICE * QUANTITY * (1 + TAXRATE)", &ctx);
         printf("  Total with tax: %.2f\n", total);
     }
 
@@ -61,7 +61,7 @@ int main() {
             .mapping_count = 0
         };
 
-        float test_cases[][2] = {
+        double test_cases[][2] = {
             {3.0, 4.0},
             {5.0, 12.0},
             {8.0, 15.0}
@@ -69,7 +69,7 @@ int main() {
 
         for (int i = 0; i < 3; i++) {
             ctx.values = test_cases[i];
-            float dist = parse_expression_with_vars(distance_formula, &ctx);
+            double dist = parse_expression_with_vars(distance_formula, &ctx);
             printf("  Distance from (%.0f, %.0f): %.2f\n",
                    test_cases[i][0], test_cases[i][1], dist);
         }
@@ -85,7 +85,7 @@ int main() {
             float time;
         } physics = {0.0, 10.0, -9.81, 2.0};
 
-        float values[] = {
+        double values[] = {
             physics.position,
             physics.velocity,
             physics.acceleration,
@@ -107,11 +107,11 @@ int main() {
         };
 
         /* Position formula: x = x0 + v0*t + 0.5*a*t^2 */
-        float position = parse_expression_with_vars("X0 + V0*T + 0.5*A*T^2", &ctx);
+        double position = parse_expression_with_vars("X0 + V0*T + 0.5*A*T^2", &ctx);
         printf("  Position after %.1fs: %.2f m\n", physics.time, position);
 
         /* Velocity formula: v = v0 + a*t */
-        float velocity = parse_expression_with_vars("V0 + A*T", &ctx);
+        double velocity = parse_expression_with_vars("V0 + A*T", &ctx);
         printf("  Velocity after %.1fs: %.2f m/s\n", physics.time, velocity);
     }
